Tiny refactor to method names and moving the mutation to its own file
